Move Token.ts file
